+++ sdk/monitor/azquery/custom_client.go
@@ -10,9 +10,12 @@ package azquery
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
+	"reflect"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 )
@@ -27,27 +30,55 @@ type LogsClientOptions struct {
 	azcore.ClientOptions
 }
 
+// LogsClient contains the methods for the LogsClient group.
+// Don't use this type directly, use NewLogsClient() instead.
+type LogsClient struct {
+	host string
+	pl   runtime.Pipeline
+}
+
+// MetricsClient contains the methods for the Metrics group.
+// Don't use this type directly, use NewMetricsClient() instead.
+type MetricsClient struct {
+	host string
+	pl   runtime.Pipeline
+}
+
 // NewLogsClient creates a client that accesses Azure Monitor logs data.
-func NewLogsClient(credential azcore.TokenCredential, options *LogsClientOptions) *LogsClient {
+func NewLogsClient(credential azcore.TokenCredential, options *LogsClientOptions) (*LogsClient, error) {
 	if options == nil {
 		options = &LogsClientOptions{}
 	}
-	authPolicy := runtime.NewBearerTokenPolicy(credential, []string{"https://api.loganalytics.io/.default"}, nil)
+	if reflect.ValueOf(options.Cloud).IsZero() {
+		options.Cloud = cloud.AzurePublic
+	}
+	c, ok := options.Cloud.Services[ServiceNameLogs]
+	if !ok || c.Audience == "" || c.Endpoint == "" {
+		return nil, errors.New("provided Cloud field is missing Azure Monitor Logs configuration")
+	}
+
+	authPolicy := runtime.NewBearerTokenPolicy(credential, []string{c.Audience + "/.default"}, nil)
 	pl := runtime.NewPipeline(moduleName, version, runtime.PipelineOptions{PerRetry: []policy.Policy{authPolicy}}, &options.ClientOptions)
-	return &LogsClient{pl: pl}
+	return &LogsClient{host: c.Endpoint, pl: pl}, nil
 }
 
 // NewMetricsClient creates a client that accesses Azure Monitor metrics data.
-func NewMetricsClient(credential azcore.TokenCredential, options *MetricsClientOptions) *MetricsClient {
+func NewMetricsClient(credential azcore.TokenCredential, options *MetricsClientOptions) (*MetricsClient, error) {
 	if options == nil {
 		options = &MetricsClientOptions{}
 	}
-	authPolicy := runtime.NewBearerTokenPolicy(credential, []string{"https://management.azure.com/.default"}, nil)
-	pl := runtime.NewPipeline(moduleName, version, runtime.PipelineOptions{PerRetry: []policy.Policy{authPolicy}}, &options.ClientOptions)
-	return &MetricsClient{pl: pl}
-}
+	if reflect.ValueOf(options.Cloud).IsZero() {
+		options.Cloud = cloud.AzurePublic
+	}
+	c, ok := options.Cloud.Services[ServiceNameMetrics]
+	if !ok || c.Audience == "" || c.Endpoint == "" {
+		return nil, errors.New("provided Cloud field is missing Azure Monitor Metrics configuration")
+	}
 
-const metricsHost string = "https://management.azure.com"
+	authPolicy := runtime.NewBearerTokenPolicy(credential, []string{c.Audience + "/.default"}, nil)
+	pl := runtime.NewPipeline(moduleName, version, runtime.PipelineOptions{PerRetry: []policy.Policy{authPolicy}}, &options.ClientOptions)
+	return &MetricsClient{host: c.Endpoint, pl: pl}, nil
+}
 
 // ErrorInfo - The code and message for an error.
 type ErrorInfo struct {